# docs/swagger/swagger.yaml
openapi: 3.0.0
servers:
  - url: "https://anke-to.trap.jp/api"
    description: production
  - url: "http://localhost:1323"
    description: local
info:
  title: anke-to API
  version: 1.0.0-oas3
  description: anke-to API
  contact:
    name: traP
    url: "https://github.com/traPtitech/anke-to"
security:
  - application:
      - read
      - write
...
paths: # TODO 変数の命名を確認する
...
  /responses/{responseID}:
    get:
      operationId: getResponse
      tags:
        - response
      description: 回答を取得します。
      parameters:
        - $ref: "#/components/parameters/responseIDInPath"
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          description: responseIDが無効です
        "403":
          description: 回答を閲覧する権限がありません。
        "404":
          description: 回答が存在しません
        "500":
          description: 回答を正常に取得できませんでした
    patch:
      operationId: editResponse
      tags:
        - response
      description: 回答を変更します。
      parameters:
        - $ref: "#/components/parameters/responseIDInPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Response"
      responses:
        "200":
          description: 正常に回答を変更できました
        "400":
          description: 与えられた回答の情報が異なります
        "403":
          description: 回答を変更する権限がありません
        "404":
          description: アンケートの回答の期限がきれたため回答が存在しません
        "405":
          description: 回答期限が過ぎたため回答できません
        "500":
          description: responseIDを取得できませんでした
    delete:
      operationId: deleteResponse
      tags:
        - response
      description: 回答を削除します
      parameters:
        - $ref: "#/components/parameters/responseIDInPath"
      responses:
        "200":
          description: 正常に回答を削除できました
        "400":
          description: 与えられた回答の情報が異なります
        "403":
          description: 回答を削除する権限がありません。
        "404":
          description: アンケートの回答の期限がきれたため回答が存在しません
        "405":
          description: 回答期限が過ぎたため回答を削除できません
        "500":
          description: responseIDを取得できませんでした
  /responses/myResponses:
    get:
      operationId: getMyResponses
      tags:
        - response
      description: 自分のすべての回答のリストを取得します。
      parameters:
        - $ref: "#/components/parameters/responseSortInQuery"
      responses:
        "200":
          description: 正常に取得できました。回答の配列を返します。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResponsesWithQuestionnaireInfo"
        "500":
          description: 自分の回答のリストを取得できませんでした
components:
  parameters:
...
    responseSortInQuery:
      name: sort
      in: query
      description:
        並び順 (作成日時が新しい "submitted_at", 作成日時が古い "-submitted_at", タイトルの昇順 "title",
        タイトルの降順 "-title", 更新日時が新しい "modified_at", 更新日時が古い
        "-modified_at" )
      schema:
        $ref: "#/components/schemas/ResponseSortType"
...
    responseIDInPath:
      name: responseID
      in: path
      required: true
      description: |
        回答ID
      schema:
        type: integer
  schemas: # TODO: description, exampleを確認する
...
    ResponseSortType:
      type: string
      description: response用のsortの種類
      enum:
        - submitted_at
        - -submitted_at
        - title
        - -title
        - modified_at
        - -modified_at
      x-enum-varnames:
        - SubmittedAtASC
        - SubmittedAtDESC
        - TitleASC
        - TitleDESC
        - ModifiedAtASC
        - ModifiedAtDESC
...
    QuestionnaireID:
      type: object
      properties:
        questionnaire_id:
          type: integer
          example: 1
      required:
        - questionnaire_id
    QuestionnaireTitle:
      type: object
      properties:
        title:
          type: string
          example: 第1回集会らん☆ぷろ募集アンケート
      required:
        - title
...
    QuestionnaireResponseDueDateTime:
      type: object
      properties:
        response_due_date_time:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00+09:00
          description: |
            回答期限。この日時を過ぎたら回答できなくなる。nullの場合は回答期限なし。
...
    QuestionnaireIsTargetingMe:
      type: object
      properties:
        is_targeting_me:
          type: boolean
          example: true
          description: |
            自分がターゲットになっているかどうか
      required:
        - is_targeting_me
    QuestionnaireCreatedAt:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00+09:00
      required:
        - created_at
    QuestionnaireModifiedAt:
      type: object
      properties:
        modified_at:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00+09:00
      required:
        - modified_at
...
    QuestionTypeText:
      type: object
      properties:
        question_type:
          type: string
          enum: [Text]
      required:
        - question_type
    QuestionTypeTextLong:
      type: object
      properties:
        question_type:
          type: string
          enum: [TextLong]
      required:
        - question_type
    QuestionTypeNumber:
      type: object
      properties:
        question_type:
          type: string
          enum: [Number]
      required:
        - question_type
    QuestionTypeSingleChoice:
      type: object
      properties:
        question_type:
          type: string
          enum: [SingleChoice]
      required:
        - question_type
    QuestionTypeMultipleChoice:
      type: object
      properties:
        question_type:
          type: string
          enum: [MultipleChoice]
      required:
        - question_type
    QuestionTypeScale:
      type: object
      properties:
        question_type:
          type: string
          enum: [Scale]
      required:
        - question_type
    NewResponse:
      type: object
      properties:
        is_draft:
          type: boolean
          example: true
        body:
          type: array
          items:
            $ref: "#/components/schemas/ResponseBody"
      required:
        - is_draft
        - body
    Response:
      allOf:
        - $ref: "#/components/schemas/QuestionnaireID"
        - type: object
          properties:
            response_id:
              type: integer
              example: 1
            respondent:
              $ref: "#/components/schemas/TraqId"
            is_anonymous:
              type: boolean
              example: true
            submitted_at:
              type: string
              format: date-time
              example: 2020-01-01T00:00:00+09:00
            modified_at:
              type: string
              format: date-time
              example: 2020-01-01T00:00:00+09:00
          required:
            - response_id
            - submitted_at
            - modified_at
        - $ref: "#/components/schemas/NewResponse"
...
    ResponsesWithQuestionnaireInfo:
      type: array
      items:
        $ref: "#/components/schemas/ResponseWithQuestionnaireInfoItem"
    QuestionnaireInfo:
      type: object
      allOf:
        - $ref: "#/components/schemas/QuestionnaireTitle"
        - $ref: "#/components/schemas/QuestionnaireResponseDueDateTime"
        - $ref: "#/components/schemas/QuestionnaireCreatedAt"
        - $ref: "#/components/schemas/QuestionnaireModifiedAt"
        - $ref: "#/components/schemas/QuestionnaireIsTargetingMe"
    ResponseWithQuestionnaireInfoItem:
      type: object
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            questionnaire_info:
              $ref: "#/components/schemas/QuestionnaireInfo"
    ResponseBody:
      oneOf:
        - $ref: "#/components/schemas/ResponseBodyText"
        - $ref: "#/components/schemas/ResponseBodyTextLong"
        - $ref: "#/components/schemas/ResponseBodyNumber"
        - $ref: "#/components/schemas/ResponseBodySingleChoice"
        - $ref: "#/components/schemas/ResponseBodyMultipleChoice"
        - $ref: "#/components/schemas/ResponseBodyScale"
    ResponseBodyText:
      allOf:
        - $ref: "#/components/schemas/QuestionTypeText"
        - $ref: "#/components/schemas/ResponseBodyBaseString"
    ResponseBodyTextLong:
      allOf:
        - $ref: "#/components/schemas/QuestionTypeTextLong"
        - $ref: "#/components/schemas/ResponseBodyBaseString"
    ResponseBodyNumber:
      allOf:
        - $ref: "#/components/schemas/QuestionTypeNumber"
        - $ref: "#/components/schemas/ResponseBodyBaseNumber"
    ResponseBodySingleChoice:
      allOf:
        - $ref: "#/components/schemas/QuestionTypeSingleChoice"
        - $ref: "#/components/schemas/ResponseBodyBaseInteger"
        - description: |
            選択肢のインデックス
    ResponseBodyMultipleChoice:
      allOf:
        - $ref: "#/components/schemas/QuestionTypeMultipleChoice"
        - type: object
          properties:
            answer:
              type: array
              items:
                type: integer
          description: |
            選択肢のインデックスの配列
          required:
            - answer
    ResponseBodyScale:
      allOf:
        - $ref: "#/components/schemas/QuestionTypeScale"
        - $ref: "#/components/schemas/ResponseBodyBaseInteger"
    ResponseBodyBaseString:
      type: object
      properties:
        answer:
          type: string
      required:
        - answer
    ResponseBodyBaseNumber:
      type: object
      properties:
        answer:
          type: number
      required:
        - answer
    ResponseBodyBaseInteger:
      type: object
      properties:
        answer:
          type: integer
      required:
        - answer
...
    TraqId:
      type: string
      example: cp20
      description: |
        traQ ID
...
  securitySchemes:
    application:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: "https://q.trap.jp/api/v3/oauth2/authorize"
          tokenUrl: "https://q.trap.jp/api/v3/oauth2/token"
          scopes:
            write: allows modifying resources
            read: allows reading resources